backend/requirements.txt
aiohttp==3.11.14
altair==5.5.0
annotated-types==0.7.0
anthropic==0.42.0
anyio==4.6.0
APScheduler==3.10.4
async-lru==2.0.4
attrs==25.1.0
beautifulsoup4==4.12.3
blinker==1.9.0
CacheControl==0.14.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
dataclasses==0.6
distro==1.9.0
fastapi==0.115.0
firebase-admin==6.5.0
gitdb==4.0.12
GitPython==3.1.44
google==3.0.0
google-ai-generativelanguage==0.6.10
google-api-core==2.20.0
google-api-python-client==2.146.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.18.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-genai==0.3.0
google-generativeai==0.8.3
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.66.1
grpcio-status==1.66.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
idna==3.10
Jinja2==3.1.5
jiter==0.5.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msgpack==1.1.0
mypy==1.14.1
mypy-extensions==1.0.0
narwhals==1.27.1
numpy==2.1.1
openai==1.47.0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
plotly==6.0.1
proto-plus==1.24.0
protobuf==5.28.2
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.9.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.23.1
rsa==4.9
ruff==0.8.4
sentry-sdk==2.19.2
setuptools==75.6.0
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
starlette==0.38.6
streamlit==1.42.2
tenacity==9.0.0
tiktoken==0.9.0
toml==0.10.2
tornado==6.4.2
tqdm==4.66.5
tenacity==9.0.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.30.6
websockets==13.1
wheel==0.45.1
